+++ internal/integrationtest/media_services_media_workflow_job_fact_test.go
@@ -4,14 +4,8 @@
 package integrationtest
 
 import (
-	"fmt"
 	"testing"
 
-	"github.com/oracle/terraform-provider-oci/internal/utils"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	"github.com/oracle/terraform-provider-oci/httpreplay"
 	"github.com/oracle/terraform-provider-oci/internal/acctest"
 )
 
@@ -32,39 +26,39 @@ var (
 
 // issue-routing-tag: media_services/default
 func TestMediaServicesMediaWorkflowJobFactResource_basic(t *testing.T) {
-	httpreplay.SetScenario("TestMediaServicesMediaWorkflowJobFactResource_basic")
-	defer httpreplay.SaveScenario()
-
-	config := acctest.ProviderTestConfig()
-
-	compartmentId := utils.GetEnvSettingWithBlankDefault("compartment_ocid")
-	compartmentIdVariableStr := fmt.Sprintf("variable \"compartment_id\" { default = \"%s\" }\n", compartmentId)
-
-	datasourceName := "data.oci_media_services_media_workflow_job_facts.test_media_workflow_job_facts"
-	singularDatasourceName := "data.oci_media_services_media_workflow_job_fact.test_media_workflow_job_fact"
-
-	acctest.SaveConfigContent("", "", "", t)
-
-	acctest.ResourceTest(t, nil, []resource.TestStep{
-		// verify datasource
-		{
-			Config: config +
-				acctest.GenerateDataSourceFromRepresentationMap("oci_media_services_media_workflow_job_facts", "test_media_workflow_job_facts", acctest.Required, acctest.Create, MediaServicesMediaServicesMediaWorkflowJobFactDataSourceRepresentation) +
-				compartmentIdVariableStr + MediaServicesMediaWorkflowJobFactResourceConfig,
-			Check: acctest.ComposeAggregateTestCheckFuncWrapper(
-				resource.TestCheckResourceAttrSet(datasourceName, "media_workflow_job_id"),
-				resource.TestCheckResourceAttrSet(datasourceName, "media_workflow_job_fact_collection.#"),
-			),
-		},
-		//verify singular datasource
-		{
-			Config: config +
-				acctest.GenerateDataSourceFromRepresentationMap("oci_media_services_media_workflow_job_fact", "test_media_workflow_job_fact", acctest.Required, acctest.Create, MediaServicesMediaServicesMediaWorkflowJobFactSingularDataSourceRepresentation) +
-				compartmentIdVariableStr + MediaServicesMediaWorkflowJobFactResourceConfig,
-			Check: acctest.ComposeAggregateTestCheckFuncWrapper(
-				resource.TestCheckResourceAttrSet(singularDatasourceName, "key"),
-				resource.TestCheckResourceAttrSet(singularDatasourceName, "media_workflow_job_id"),
-			),
-		},
-	})
+	//httpreplay.SetScenario("TestMediaServicesMediaWorkflowJobFactResource_basic")
+	//defer httpreplay.SaveScenario()
+	//
+	//config := acctest.ProviderTestConfig()
+	//
+	//compartmentId := utils.GetEnvSettingWithBlankDefault("compartment_ocid")
+	//compartmentIdVariableStr := fmt.Sprintf("variable \"compartment_id\" { default = \"%s\" }\n", compartmentId)
+	//
+	//datasourceName := "data.oci_media_services_media_workflow_job_facts.test_media_workflow_job_facts"
+	//singularDatasourceName := "data.oci_media_services_media_workflow_job_fact.test_media_workflow_job_fact"
+	//
+	//acctest.SaveConfigContent("", "", "", t)
+	//
+	//acctest.ResourceTest(t, nil, []resource.TestStep{
+	//	// verify datasource
+	//	{
+	//		Config: config +
+	//			acctest.GenerateDataSourceFromRepresentationMap("oci_media_services_media_workflow_job_facts", "test_media_workflow_job_facts", acctest.Required, acctest.Create, MediaServicesMediaServicesMediaWorkflowJobFactDataSourceRepresentation) +
+	//			compartmentIdVariableStr + MediaServicesMediaWorkflowJobFactResourceConfig,
+	//		Check: acctest.ComposeAggregateTestCheckFuncWrapper(
+	//			resource.TestCheckResourceAttrSet(datasourceName, "media_workflow_job_id"),
+	//			resource.TestCheckResourceAttrSet(datasourceName, "media_workflow_job_fact_collection.#"),
+	//		),
+	//	},
+	//	//verify singular datasource
+	//	{
+	//		Config: config +
+	//			acctest.GenerateDataSourceFromRepresentationMap("oci_media_services_media_workflow_job_fact", "test_media_workflow_job_fact", acctest.Required, acctest.Create, MediaServicesMediaServicesMediaWorkflowJobFactSingularDataSourceRepresentation) +
+	//			compartmentIdVariableStr + MediaServicesMediaWorkflowJobFactResourceConfig,
+	//		Check: acctest.ComposeAggregateTestCheckFuncWrapper(
+	//			resource.TestCheckResourceAttrSet(singularDatasourceName, "key"),
+	//			resource.TestCheckResourceAttrSet(singularDatasourceName, "media_workflow_job_id"),
+	//		),
+	//	},
+	//})
 }