+++ internal/integrationtest/media_services_stream_cdn_config_test.go
@@ -50,10 +50,12 @@ var (
 		"config":                  acctest.RepresentationGroup{RepType: acctest.Required, Group: MediaServicesStreamCdnConfigConfigRepresentation},
 		"display_name":            acctest.Representation{RepType: acctest.Required, Create: `displayName`, Update: `displayName2`},
 		"distribution_channel_id": acctest.Representation{RepType: acctest.Required, Create: `${oci_media_services_stream_distribution_channel.test_stream_distribution_channel.id}`},
+		"is_lock_override":        acctest.Representation{RepType: acctest.Required, Create: `true`, Update: `true`},
 		"defined_tags":            acctest.Representation{RepType: acctest.Optional, Create: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "value")}`, Update: `${map("${oci_identity_tag_namespace.tag-namespace1.name}.${oci_identity_tag.tag1.name}", "updatedValue")}`},
 		"freeform_tags":           acctest.Representation{RepType: acctest.Optional, Create: map[string]string{"bar-key": "value"}, Update: map[string]string{"Department": "Accounting"}},
 		"is_enabled":              acctest.Representation{RepType: acctest.Optional, Create: `false`, Update: `true`},
-		"lifecycle":               acctest.RepresentationGroup{RepType: acctest.Optional, Group: ignoreDefinedTags},
+		"locks":                   acctest.RepresentationGroup{RepType: acctest.Optional, Group: MediaServicesStreamCdnConfigLocksRepresentation},
+		"lifecycle":               acctest.RepresentationGroup{RepType: acctest.Optional, Group: ignoreDefinedTagsAndLocks},
 	}
 
 	MediaServicesStreamCdnConfigConfigRepresentation = map[string]interface{}{
@@ -71,6 +73,15 @@ var (
 		"origin_auth_sign_type":          acctest.Representation{RepType: acctest.Optional, Create: `ForwardURL`},
 	}
 
+	MediaServicesStreamCdnConfigLocksRepresentation = map[string]interface{}{
+		"compartment_id": acctest.Representation{RepType: acctest.Required, Create: `${var.compartment_id}`},
+		"type":           acctest.Representation{RepType: acctest.Required, Create: `FULL`},
+		"message":        acctest.Representation{RepType: acctest.Optional, Create: `message`},
+	}
+
+	MediaServicesStreamCdnConfigResourceDependencies = DefinedTagsDependencies +
+		acctest.GenerateResourceFromRepresentationMap("oci_media_services_stream_distribution_channel", "test_stream_distribution_channel", acctest.Required, acctest.Create, MediaServicesStreamDistributionChannelRepresentation)
+
 	MediaServicesStreamCdnConfigRepresentationWithEdge = map[string]interface{}{
 		"config":                  acctest.RepresentationGroup{RepType: acctest.Required, Group: MediaServicesStreamCdnConfigConfigRepresentationWithEdge},
 		"display_name":            acctest.Representation{RepType: acctest.Required, Create: `displayName`, Update: `displayName2`},
@@ -79,12 +90,10 @@ var (
 		"freeform_tags":           acctest.Representation{RepType: acctest.Optional, Create: map[string]string{"bar-key": "value"}, Update: map[string]string{"Department": "Accounting"}},
 		"is_enabled":              acctest.Representation{RepType: acctest.Optional, Create: `false`, Update: `true`},
 	}
+
 	MediaServicesStreamCdnConfigConfigRepresentationWithEdge = map[string]interface{}{
 		"type": acctest.Representation{RepType: acctest.Required, Create: `EDGE`, Update: `EDGE`},
 	}
-
-	MediaServicesStreamCdnConfigResourceDependencies = acctest.GenerateResourceFromRepresentationMap("oci_media_services_stream_distribution_channel", "test_stream_distribution_channel", acctest.Required, acctest.Create, MediaServicesStreamDistributionChannelRepresentation) +
-		DefinedTagsDependencies
 )
 
 // issue-routing-tag: media_services/default
@@ -248,11 +257,13 @@ func TestMediaServicesStreamCdnConfigResource_basic(t *testing.T) {
 		},
 		// verify resource import
 		{
-			Config:                  config + MediaServicesStreamCdnConfigRequiredOnlyResource,
-			ImportState:             true,
-			ImportStateVerify:       true,
-			ImportStateVerifyIgnore: []string{},
-			ResourceName:            resourceName,
+			Config:            config + MediaServicesStreamCdnConfigRequiredOnlyResource,
+			ImportState:       true,
+			ImportStateVerify: true,
+			ImportStateVerifyIgnore: []string{
+				"is_lock_override",
+			},
+			ResourceName: resourceName,
 		},
 	})
 }