+++ internal/service/database_migration/register_datasource.go
@@ -6,9 +6,9 @@ package database_migration
 import "github.com/oracle/terraform-provider-oci/internal/tfresource"
 
 func RegisterDatasource() {
-	tfresource.RegisterDatasource("oci_database_migration_agent", DatabaseMigrationAgentDataSource())
-	tfresource.RegisterDatasource("oci_database_migration_agent_images", DatabaseMigrationAgentImagesDataSource())
-	tfresource.RegisterDatasource("oci_database_migration_agents", DatabaseMigrationAgentsDataSource())
+	//tfresource.RegisterDatasource("oci_database_migration_agent", DatabaseMigrationAgentDataSource())
+	//tfresource.RegisterDatasource("oci_database_migration_agent_images", DatabaseMigrationAgentImagesDataSource())
+	//tfresource.RegisterDatasource("oci_database_migration_agents", DatabaseMigrationAgentsDataSource())
 	tfresource.RegisterDatasource("oci_database_migration_connection", DatabaseMigrationConnectionDataSource())
 	tfresource.RegisterDatasource("oci_database_migration_connections", DatabaseMigrationConnectionsDataSource())
 	tfresource.RegisterDatasource("oci_database_migration_job", DatabaseMigrationJobDataSource())