+++ internal/service/management_agent/register_resource.go
@@ -7,5 +7,6 @@ import "github.com/oracle/terraform-provider-oci/internal/tfresource"
 
 func RegisterResource() {
 	tfresource.RegisterResource("oci_management_agent_management_agent", ManagementAgentManagementAgentResource())
+	tfresource.RegisterResource("oci_management_agent_management_agent_data_source", ManagementAgentManagementAgentDataSourceResource())
 	tfresource.RegisterResource("oci_management_agent_management_agent_install_key", ManagementAgentManagementAgentInstallKeyResource())
 }