+++ internal/service/mysql/mysql_mysql_db_system_data_source.go
@@ -184,6 +184,12 @@ func (s *MysqlMysqlDbSystemDataSourceCrud) SetData() error {
 		s.D.Set("port_x", *s.Res.PortX)
 	}
 
+	if s.Res.SecureConnections != nil {
+		s.D.Set("secure_connections", []interface{}{SecureConnectionDetailsToMap(s.Res.SecureConnections)})
+	} else {
+		s.D.Set("secure_connections", nil)
+	}
+
 	if s.Res.ShapeName != nil {
 		s.D.Set("shape_name", *s.Res.ShapeName)
 	}